Move all the relevant bits to test_collection
